backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.3
celery==5.3.6
neo4j==5.19.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0